Unit 2 - Lesson 2 - Part 4 - Adding a UI
Did a lot of testing this weekend and it indeed seems to be related to
the collision system. Starting at line 34 in systems/collisions, there
is a var sizeB=entity,components.collision.size that is trying to find
a size for entityB, which is the pipe. The pipe does not have a size,
only a width and a height.